{{cookiecutter.project_slug}}/requirements.txt
alembic==1.0.10
aniso8601==6.0.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
Click==7.0
coverage==4.5.3
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.4.0
Flask-Cors==3.0.8
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Mako==1.0.9
MarkupSafe==1.1.1
pycparser==2.19
psycopg2-binary==2.7.6.1
PyJWT==1.4.2
pyrsistent==0.15.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
PyYAML==5.1
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.15.2
wincertstore==0.2
gunicorn==19.9.0
supervisord==4.1.0
